Move template scripts to ui folder
