+++ app/static/js/components/ui/template_menu.js
@@ -1,4 +1,4 @@
-import { TemplateModal } from "../components/modal.js";
+import { TemplateModal } from "../modal.js";
 
 
 export class TemplateMenu{
