+++ app/static/js/pages/timeline/timeline.js
@@ -2,7 +2,7 @@ import { checkLeftMarker } from '../../components/timeline/timeline_left_markers
 import { adjustEpochLine } from '../../components/timeline/epoch_line_adjust.js';
 import { marginFix } from '../../components/timeline/margin_fix.js';
 import { Modal } from '../../components/modal.js';
-import { TimelineFooter } from '../../components/footer.js';
+import { TimelineFooter } from '../../components/ui/footer.js';
 
 // Offset epoch elements
 adjustEpochLine();
